+++ assets/js/state/requests.ts
@@ -1,4 +1,4 @@
-import { get } from "../fetch";
+import { get } from "../util/fetch";
 import { zip } from "../unused/track/_utils";
