--- assets/js/track.ts
@@ -1,10 +1,0 @@
-// Entrypoint for track module
-
-export { VariantTrack } from "./unused/track/_variant";
-export { TranscriptTrack } from "./unused/track/_transcript";
-// export { AnnotationTrack } from "./track/annotation";
-export { BaseScatterTrack } from "./unused/track/_base";
-export {
-  CytogeneticIdeogram,
-  setupGenericEventManager,
-} from "./unused/track/_ideogram";
